+++ Sources/Focus.swift
@@ -1,9 +1,19 @@
-public struct Focus {
-
-    public static var failureHandler: ((_ message: String, _ file: StaticString, _ line: UInt) -> (Void))?
-    public static var successHandler: ((_ file: StaticString, _ line: UInt) -> (Void))?
+struct Focus {
+    
+    /**
+     Responsible for handling pass and fail.
+     */
+    static var reporter: Reportable? = defaultReporter()
 }
 
 public func expect<Item>(_ item: Item) -> Expect<Item> {
     return Expect(item: item)
+}
+
+public func set(reporter: Reportable) {
+    Focus.reporter = reporter
+}
+
+public func defaultReporter() -> DefaultReporter {
+    return DefaultReporter.sharedReporter
 }