+++ Tests/FocusTests/WhenCreatingAnExpectObject.swift
@@ -10,22 +10,14 @@ import XCTest
 
 class WhenCreatingAnExpectObject: XCTestCase {
     
-    var failureData:(used: Bool, comment: String, file: String, line: UInt) = (false, "", "", 0)
-    var successData:(used: Bool, file: String, line: UInt) = (false, "", 0)
+    var reporter: Reporter!
     
     override func setUp() {
         super.setUp()
         // Put setup code here. This method is called before the invocation of each test method in the class.
-        failureData = (false, "", "", 0)
-        successData = (false, "", 0)
-        
-        Focus.failureHandler = { message, file, line in
-            self.failureData = (used: true, comment: message, file: String(describing: file), line: line)
-        }
-        
-        Focus.successHandler = { file, line in
-            self.successData = (used: true, file: String(describing: file), line: line)
-        }
+        reporter = Reporter()
+        reporter.resetData()
+        Focus.reporter = reporter
     }
     
     override func tearDown() {
@@ -47,8 +39,8 @@ class WhenCreatingAnExpectObject: XCTestCase {
         let expectedLine: UInt = #line-2
         let expectedFile = #file
         
-        XCTAssertEqual(successData.file, expectedFile)
-        XCTAssertEqual(successData.line, expectedLine)
+        XCTAssertEqual(reporter.successData.file, expectedFile)
+        XCTAssertEqual(reporter.successData.line, expectedLine)
     }
     
     func test_expect_UsesCommentFileAndLineNumberWhenFailes() {
@@ -59,9 +51,9 @@ class WhenCreatingAnExpectObject: XCTestCase {
         let expectedLine: UInt = #line-2
         let expectedFile = #file
         
-        XCTAssertEqual(failureData.file, expectedFile)
-        XCTAssertEqual(failureData.line, expectedLine)
-        XCTAssertEqual(failureData.comment, comment)
+        XCTAssertEqual(reporter.failureData.file, expectedFile)
+        XCTAssertEqual(reporter.failureData.line, expectedLine)
+        XCTAssertEqual(reporter.failureData.comment, comment)
     }
     
     func test_expect_UsesCommentFileAndLineNumberWhenFailesBecauseNonBoolIsPassedIn() {
@@ -72,9 +64,9 @@ class WhenCreatingAnExpectObject: XCTestCase {
         let expectedLine: UInt = #line-2
         let expectedFile = #file
         
-        XCTAssertEqual(failureData.file, expectedFile)
-        XCTAssertEqual(failureData.line, expectedLine)
-        XCTAssertEqual(failureData.comment, comment)
+        XCTAssertEqual(reporter.failureData.file, expectedFile)
+        XCTAssertEqual(reporter.failureData.line, expectedLine)
+        XCTAssertEqual(reporter.failureData.comment, comment)
     }
 
 }