+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 
+## [0.4.0] - 2017-03-27
+### Added
+- A reporter to be able to customize test output and hanldle passing and failing tests.
+
+### Changed
+- Remove old way of handling failure and success messages.
+
+
 ## [0.3.0] - 2017-02-08
 ### Added